[AIRFLOW-6162] Add back serialization as a module (#6718)
(cherry-picked from 0abefe9243388e45d676508b9f1cb84d1c8d3d8b)
