+++ airflow/models/serialized_dag.py
@@ -29,7 +29,7 @@ from sqlalchemy.sql import exists
 
 from airflow import DAG
 from airflow.models.base import ID_LEN, Base
-from airflow.serialization.serialized_dag import SerializedDAG
+from airflow.serialization.serialized_objects import SerializedDAG
 from airflow.settings import json
 from airflow.utils import db, timezone
 from airflow.utils.log.logging_mixin import LoggingMixin
