+++ airflow/serialization/serialized_objects.py
@@ -17,7 +17,7 @@
 # specific language governing permissions and limitations
 # under the License.
 
-"""Utils for DAG serialization with JSON."""
+"""Serialized DAG and BaseOperator"""
 
 import datetime
 import enum
@@ -28,24 +28,21 @@ from typing import TYPE_CHECKING, Optional, Union, Dict
 import pendulum
 from dateutil import relativedelta
 
-from airflow.exceptions import AirflowException
-from airflow.models import DAG
+from airflow import DAG, AirflowException, LoggingMixin
 from airflow.models.baseoperator import BaseOperator
 from airflow.models.connection import Connection
 from airflow.serialization.enums import DagAttributeTypes as DAT, Encoding
-from airflow.serialization.json_schema import Validator
+from airflow.serialization.json_schema import Validator, load_dag_schema
 from airflow.settings import json
-from airflow.utils.log.logging_mixin import LoggingMixin
 from airflow.www.utils import get_python_source
+
+try:
+    from inspect import signature
+except ImportError:
+    from funcsigs import signature  # type: ignore
 
 if TYPE_CHECKING:
     from inspect import Parameter
-
-
-LOG = LoggingMixin().log
-
-# Serialization failure returns 'failed'.
-FAILED = 'serialization_failed'
 
 
 class BaseSerialization:
@@ -59,7 +56,6 @@ class BaseSerialization:
     _datetime_types = (datetime.datetime,)
 
     # Object types that are always excluded in serialization.
-    # FIXME: not needed if _included_fields of DAG and operator are customized.
     _excluded_types = (logging.Logger, Connection, type)
 
     _json_schema = None     # type: Optional[Validator]
@@ -167,8 +163,6 @@ class BaseSerialization:
         (3) Operator has a special field CLASS to record the original class
             name for displaying in UI.
         """
-        from airflow.serialization.serialized_dag import SerializedDAG
-        from airflow.serialization.serialized_baseoperator import SerializedBaseOperator
         try:
             if cls._is_primitive(var):
                 # enum.IntEnum is an int instance, it causes json dumps error so we use its value.
@@ -219,8 +213,6 @@ class BaseSerialization:
     @classmethod
     def _deserialize(cls, encoded_var):  # pylint: disable=too-many-return-statements
         """Helper function of depth first search for deserialization."""
-        from airflow.serialization.serialized_dag import SerializedDAG
-        from airflow.serialization.serialized_baseoperator import SerializedBaseOperator
         # JSON primitives (except for dict) are not encoded.
         if cls._is_primitive(encoded_var):
             return encoded_var
@@ -278,3 +270,234 @@ class BaseSerialization:
         if attrname in cls._CONSTRUCTOR_PARAMS and cls._CONSTRUCTOR_PARAMS[attrname].default is value:
             return True
         return False
+
+
+class SerializedBaseOperator(BaseOperator, BaseSerialization):
+    """A JSON serializable representation of operator.
+
+    All operators are casted to SerializedBaseOperator after deserialization.
+    Class specific attributes used by UI are move to object attributes.
+    """
+    _decorated_fields = {'executor_config', }
+
+    _CONSTRUCTOR_PARAMS = {
+        k: v for k, v in signature(BaseOperator).parameters.items()
+        if v.default is not v.empty and v.default is not None
+    }
+
+    def __init__(self, *args, **kwargs):
+        super(SerializedBaseOperator, self).__init__(*args, **kwargs)
+        # task_type is used by UI to display the correct class type, because UI only
+        # receives BaseOperator from deserialized DAGs.
+        self._task_type = 'BaseOperator'
+        # Move class attributes into object attributes.
+        self.ui_color = BaseOperator.ui_color
+        self.ui_fgcolor = BaseOperator.ui_fgcolor
+        self.template_fields = BaseOperator.template_fields
+        # subdag parameter is only set for SubDagOperator.
+        # Setting it to None by default as other Operators do not have that field
+        self.subdag = None
+        self.operator_extra_links = BaseOperator.operator_extra_links
+
+    @property
+    def task_type(self):
+        # type: () -> str
+        # Overwrites task_type of BaseOperator to use _task_type instead of
+        # __class__.__name__.
+        return self._task_type
+
+    @task_type.setter
+    def task_type(self, task_type):
+        # type: (str) -> None
+        self._task_type = task_type
+
+    @classmethod
+    def serialize_operator(cls, op):
+        # type: (BaseOperator) -> dict
+        """Serializes operator into a JSON object.
+        """
+        serialize_op = cls.serialize_to_json(op, cls._decorated_fields)
+        # Adds a new task_type field to record the original operator class.
+        serialize_op['_task_type'] = op.__class__.__name__
+        serialize_op['_task_module'] = op.__class__.__module__
+        return serialize_op
+
+    @classmethod
+    def deserialize_operator(cls, encoded_op):
+        # type: (dict) -> BaseOperator
+        """Deserializes an operator from a JSON object.
+        """
+        from airflow.plugins_manager import operator_extra_links
+
+        op = SerializedBaseOperator(task_id=encoded_op['task_id'])
+
+        # Extra Operator Links
+        op_extra_links_from_plugin = {}
+
+        for ope in operator_extra_links:
+            for operator in ope.operators:
+                if operator.__name__ == encoded_op["_task_type"] and \
+                        operator.__module__ == encoded_op["_task_module"]:
+                    op_extra_links_from_plugin.update({ope.name: ope})
+
+        setattr(op, "operator_extra_links", list(op_extra_links_from_plugin.values()))
+
+        for k, v in encoded_op.items():
+
+            if k == "_downstream_task_ids":
+                v = set(v)
+            elif k == "subdag":
+                v = SerializedDAG.deserialize_dag(v)
+            elif k in {"retry_delay", "execution_timeout"}:
+                v = cls._deserialize_timedelta(v)
+            elif k.endswith("_date"):
+                v = cls._deserialize_datetime(v)
+            elif k in cls._decorated_fields or k not in op.get_serialized_fields():
+                v = cls._deserialize(v)
+            # else use v as it is
+
+            setattr(op, k, v)
+
+        for k in op.get_serialized_fields() - set(encoded_op.keys()) - set(
+                cls._CONSTRUCTOR_PARAMS.keys()):
+            setattr(op, k, None)
+
+        return op
+
+    @classmethod
+    def _is_excluded(cls, var, attrname, op):
+        if var is not None and op.has_dag() and attrname.endswith("_date"):
+            # If this date is the same as the matching field in the dag, then
+            # don't store it again at the task level.
+            dag_date = getattr(op.dag, attrname, None)
+            if var is dag_date or var == dag_date:
+                return True
+        if attrname in {"executor_config", "params"} and not var:
+            # Don't store empty executor config or params dicts.
+            return True
+        return super(SerializedBaseOperator, cls)._is_excluded(var, attrname, op)
+
+
+class SerializedDAG(DAG, BaseSerialization):
+    """A JSON serializable representation of DAG.
+
+    A stringified DAG can only be used in the scope of scheduler and webserver, because fields
+    that are not serializable, such as functions and customer defined classes, are casted to
+    strings.
+
+    Compared with SimpleDAG: SerializedDAG contains all information for webserver.
+    Compared with DagPickle: DagPickle contains all information for worker, but some DAGs are
+    not pickle-able. SerializedDAG works for all DAGs.
+    """
+
+    _decorated_fields = {'schedule_interval', 'default_args'}
+
+    @staticmethod
+    def __get_constructor_defaults():  # pylint: disable=no-method-argument
+        param_to_attr = {
+            'concurrency': '_concurrency',
+            'description': '_description',
+            'default_view': '_default_view',
+            'access_control': '_access_control',
+        }
+        return {
+            param_to_attr.get(k, k): v for k, v in signature(DAG).parameters.items()
+            if v.default is not v.empty and v.default is not None
+        }
+
+    _CONSTRUCTOR_PARAMS = __get_constructor_defaults.__func__()  # type: ignore
+    del __get_constructor_defaults
+
+    _json_schema = load_dag_schema()
+
+    @classmethod
+    def serialize_dag(cls, dag):
+        # type: (DAG) -> dict
+        """Serializes a DAG into a JSON object.
+        """
+        serialize_dag = cls.serialize_to_json(dag, cls._decorated_fields)
+
+        serialize_dag["tasks"] = [cls._serialize(task) for _, task in dag.task_dict.items()]
+        return serialize_dag
+
+    @classmethod
+    def deserialize_dag(cls, encoded_dag):
+        # type: (dict) -> 'SerializedDAG'
+        """Deserializes a DAG from a JSON object.
+        """
+
+        dag = SerializedDAG(dag_id=encoded_dag['_dag_id'])
+
+        for k, v in encoded_dag.items():
+            if k == "_downstream_task_ids":
+                v = set(v)
+            elif k == "tasks":
+                v = {
+                    task["task_id"]: SerializedBaseOperator.deserialize_operator(task) for task in v
+                }
+                k = "task_dict"
+            elif k == "timezone":
+                v = cls._deserialize_timezone(v)
+            elif k in {"retry_delay", "execution_timeout"}:
+                v = cls._deserialize_timedelta(v)
+            elif k.endswith("_date"):
+                v = cls._deserialize_datetime(v)
+            elif k in cls._decorated_fields:
+                v = cls._deserialize(v)
+            # else use v as it is
+
+            setattr(dag, k, v)
+
+        keys_to_set_none = dag.get_serialized_fields() - set(encoded_dag.keys()) - set(
+            cls._CONSTRUCTOR_PARAMS.keys())
+        for k in keys_to_set_none:
+            setattr(dag, k, None)
+
+        setattr(dag, 'full_filepath', dag.fileloc)
+        for task in dag.task_dict.values():
+            task.dag = dag
+            serializable_task = task  # type: BaseOperator
+
+            for date_attr in ["start_date", "end_date"]:
+                if getattr(serializable_task, date_attr) is None:
+                    setattr(serializable_task, date_attr, getattr(dag, date_attr))
+
+            if serializable_task.subdag is not None:
+                setattr(serializable_task.subdag, 'parent_dag', dag)
+                serializable_task.subdag.is_subdag = True
+
+            for task_id in serializable_task.downstream_task_ids:
+                # Bypass set_upstream etc here - it does more than we want
+                # noinspection PyProtectedMember
+                dag.task_dict[task_id]._upstream_task_ids.add(task_id)  # pylint: disable=protected-access
+
+        return dag
+
+    @classmethod
+    def to_dict(cls, var):
+        # type: (...) -> dict
+        """Stringifies DAGs and operators contained by var and returns a dict of var.
+        """
+        json_dict = {
+            "__version": cls.SERIALIZER_VERSION,
+            "dag": cls.serialize_dag(var)
+        }
+
+        # Validate Serialized DAG with Json Schema. Raises Error if it mismatches
+        cls.validate_schema(json_dict)
+        return json_dict
+
+    @classmethod
+    def from_dict(cls, serialized_obj):
+        # type: (dict) -> 'SerializedDAG'
+        """Deserializes a python dict in to the DAG and operators it contains."""
+        ver = serialized_obj.get('__version', '<not present>')
+        if ver != cls.SERIALIZER_VERSION:
+            raise ValueError("Unsure how to deserialize version {!r}".format(ver))
+        return cls.deserialize_dag(serialized_obj['dag'])
+
+
+LOG = LoggingMixin().log
+
+# Serialization failure returns 'failed'.
+FAILED = 'serialization_failed'
